requirements.txt
fastapi
uvicorn
openai
faiss-cpu
beautifulsoup4
requests
numpy
